IMPALA-10044: Fix cleanup for bootstrap_toolchain.py failure case
If DownloadUnpackTarball::download()'s wget_and_unpack_package call
hits an exception, the exception handler cleans up any created
directories. Currently, it erroneously cleans up the directory where
the tarballs are downloaded even when it is not a temporary directory.
This would delete the entire toolchain.

This fixes the cleanup to only delete that directory if it is a
temporary directory.

Testing:
 - Simulated exception from wget_and_unpack_package and verified
   behavior.

Reviewed-on: http://gerrit.cloudera.org:8080/16294
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>

+++ bin/bootstrap_toolchain.py
@@ -182,7 +182,8 @@ class DownloadUnpackTarball(object):
       # Clean up any partially-unpacked result.
       if os.path.isdir(unpack_dir):
         shutil.rmtree(unpack_dir)
-      if os.path.isdir(download_dir):
+      # Only delete the download directory if it is a temporary directory
+      if download_dir != self.destination_basedir and os.path.isdir(download_dir):
         shutil.rmtree(download_dir)
       raise
     if self.makedir:
